# apps/web/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditRow": {
        "description": "Ein Eintrag, wie ihn ein Mensch liest.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "entry_hash": {
            "title": "Entry Hash",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          }
        },
        "required": [
          "id",
          "occurred_at",
          "actor",
          "action",
          "resource",
          "details",
          "entry_hash"
        ],
        "title": "AuditRow",
        "type": "object"
      },
...
      "ChainStatus": {
        "description": "Das Ergebnis der Kettenprüfung.",
        "properties": {
          "breaks": {
            "items": {
              "type": "string"
            },
            "title": "Breaks",
            "type": "array"
          },
          "checked": {
            "title": "Checked",
            "type": "integer"
          },
          "intact": {
            "title": "Intact",
            "type": "boolean"
          }
        },
        "required": [
          "intact",
          "checked",
          "breaks"
        ],
        "title": "ChainStatus",
        "type": "object"
      },
      "GrantView": {
        "properties": {
          "constraints": {
            "additionalProperties": true,
            "title": "Constraints",
            "type": "object"
          },
          "expired": {
            "title": "Expired",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/PermissionMode"
          }
        },
        "required": [
          "mode",
          "constraints",
          "granted_at",
          "expired"
        ],
        "title": "GrantView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PermissionMode": {
        "description": "Wie eine Berechtigung erteilt ist.",
        "enum": [
          "deny",
          "confirm",
          "allow"
        ],
        "title": "PermissionMode",
        "type": "string"
      },
...
      "SetGrantRequest": {
        "description": "Was ein Nutzer über einen Scope entscheidet.\n\nKein ``scope``-Feld: Der Scope steht im Pfad. Zwei Angaben desselben\nGegenstands sind eine Gelegenheit, sie auseinanderlaufen zu lassen.",
        "properties": {
          "constraints": {
            "additionalProperties": true,
            "title": "Constraints",
            "type": "object"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "mode": {
            "$ref": "#/components/schemas/PermissionMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "SetGrantRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Alle Endpunkte außer der Erstinbetriebnahme und dem Anmelde-Handshake verlangen eine gültige Sitzung. Die Identität stammt ausschließlich aus dieser Sitzung, nie aus Angaben im Request.",
    "summary": "Persönliches KI-Assistenzsystem",
    "title": "JARVIS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/audit": {
      "get": {
        "description": "Die jüngsten Einträge, neueste zuerst — das Aktivitätsprotokoll.\n\nZeigt **eigene** Einträge. Anders als die Kettenprüfung, die das ganze\nSystem betrifft: Hier geht es um Auskunft an einen Menschen darüber, was in\nseinem Namen geschehen ist.",
        "operationId": "list_audit_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditRow"
                  },
                  "title": "Response List Audit Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit",
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/verify": {
      "get": {
        "description": "Rechnet die Kette nach.\n\n**Ohne ``limit`` wird alles geprüft**, und nur das beantwortet die Frage\n„ist irgendetwas verändert worden?\". Ein Ausschnitt beantwortet „ist seit\nEintrag N etwas verändert worden?\" — eine andere und schwächere Frage, die\nhier trotzdem angeboten wird, weil ein wachsendes Protokoll sonst irgendwann\nnur noch nachts prüfbar wäre.\n\n**Warum die Antwort nicht nach Nutzern getrennt ist.** Die Kette ist eine\nEigenschaft des Systems, nicht einer Person: Ein fremder Eintrag dazwischen\ngehört zur Kette und wird mitgerechnet. Dieses System hat einen Nutzer; wenn\nes mehrere bekommt, gehört diese Auskunft an eine Rolle und nicht an jede\nSitzung.",
        "operationId": "verify_chain_endpoint_audit_verify_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Chain Endpoint",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/permissions/{scope}": {
      "delete": {
        "description": "Zieht eine Berechtigung zurück.\n\n``204`` auch dann, wenn es nichts zurückzuziehen gab: Danach gilt in beiden\nFällen dasselbe — der Nutzer hat dieses Recht nicht erteilt. Ein ``404``\nwäre eine Auskunft darüber, was jemand erteilt hat, und die gehört nicht in\neinen Statuscode.\n\n**Sofort wirksam.** Die Policy Engine liest bei jedem Aufruf neu; ein Lauf,\nder gerade zwischen zwei Schritten steht, findet das Recht beim nächsten\nnicht mehr vor.",
        "operationId": "revoke_permission_permissions__scope__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Permission",
        "tags": [
          "permissions"
        ]
      },
      "put": {
        "description": "Erteilt, ändert oder verweigert einen Scope — vollständig.\n\nEin ``PUT`` und kein ``PATCH``, und das ist bedeutungstragend: Was hier\nsteht, gilt danach. Ein Zusammenführen mit dem alten Stand wäre die Art von\nBequemlichkeit, bei der eine Pfadgrenze aus der Vorwoche eine neue\nErteilung still erweitert.",
        "operationId": "set_permission_permissions__scope__put",
        "parameters": [
          {
            "in": "path",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Permission",
        "tags": [
          "permissions"
        ]
      }
    },
...
  }
}